feat: make reusable gridContainer component
