+++ src/components/GridContainer.tsx
@@ -1,33 +1,17 @@
 "use client";
 
-import dynamic from "next/dynamic";
+import { ReactNode } from "react";
 import { Layouts, Responsive, WidthProvider } from "react-grid-layout";
-
-import { Card } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-
-import { Project } from "@/data/project-dto";
-import { OwnerData } from "@/types/ownerData";
-import ProjectCard from "./ProjectCard";
-import GithubCard from "./GithubCard";
-import ThemeToggle from "./ThemeToggle";
-import { useLayoutsContext } from "./LayoutsContext";
 
 const ResponsiveGridLayout = WidthProvider(Responsive);
 
-const DynamicMap = dynamic(() => import("@/components/Map"), {
-  loading: () => <Skeleton className="h-full w-full" />,
-  ssr: false,
-});
-
 export default function GridContainer({
-  ownerData,
-  projects,
+  children,
+  layouts,
 }: {
-  ownerData: OwnerData | null;
-  projects: Project[];
+  children: ReactNode;
+  layouts: Layouts;
 }) {
-  const { layouts } = useLayoutsContext();
   return (
     <ResponsiveGridLayout
       draggableCancel=".cancelDrag"
@@ -50,26 +34,7 @@ export default function GridContainer({
       }}
       // isBounded={true}
     >
-      <Card variant="item" key="me">
-        <p>{ownerData?.aboutMe}</p>
-      </Card>
-      <Card variant="item" key="toggle-theme">
-        <ThemeToggle />
-      </Card>
-      <Card variant="item" key="maps" className="overflow-hidden">
-        <DynamicMap />
-      </Card>
-      <Card variant="item" key="social-links">
-        <GithubCard githubUser={ownerData?.githubUser} />
-      </Card>
-      {projects.map((project, index) => (
-        <Card variant="item" key={project.slug} className="overflow-hidden">
-          <ProjectCard
-            project={project}
-            isHorizontal={index === 1 ? true : false}
-          />
-        </Card>
-      ))}
+      {children}
     </ResponsiveGridLayout>
   );
 }
